feat(stats) separate featch function from hookm
