+++ src/components/Dashboard/StatsSection/StatsOverview/hooks/useStatsCards.hook.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { FaTrophy, FaFire, FaClock, FaStar } from "react-icons/fa";
 import { UserStats } from "../../../../../types/stats.types";
-import { useFormatTime } from './useFormatTime';
+import { useFormatTime } from './useFormatTime.hook';
 
 export const useStatsCards = (stats: UserStats | undefined) => {
   const formatTime = useFormatTime();
